[refactor] divide packages - common - domain - external
