+++ src/main/java/park/brothers/runwith_back/domain/Login/LoginController.java
@@ -1,4 +1,4 @@
-package park.brothers.runwith_back.Login;
+package park.brothers.runwith_back.domain.Login;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
@@ -11,7 +11,7 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import park.brothers.runwith_back.Runner.Runner;
+import park.brothers.runwith_back.domain.Runner.Runner;
 
 import java.util.HashMap;
 import java.util.Map;
